+++ mobile/src/main/java/me/calebjones/spacelaunchnow/content/database/ListPreferences.java
@@ -18,6 +18,7 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.Locale;
 import java.util.TimeZone;
+import java.util.concurrent.TimeUnit;
 
 import timber.log.Timber;