+++ mobile/src/main/java/me/calebjones/spacelaunchnow/utils/Connectivity.java
@@ -26,8 +26,8 @@ public class Connectivity {
     public static String getNetworkStatus(Context context){
         ConnectivityManager cm = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
         if (cm != null && cm.getActiveNetworkInfo() != null){
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                return cm.getActiveNetwork().toString();
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M && cm.getActiveNetworkInfo().getTypeName() != null) {
+                return cm.getActiveNetworkInfo().getTypeName();
             }
         }
         return "Unknown";